# [WIP] Performance optimization for dot(csr, rsp) on cpu