--- a/src/operator/tensor/dot-inl.h
+++ b/src/operator/tensor/dot-inl.h
@@ -444,23 +444,16 @@ struct DotCsrRspDnsByRowBlocks {
                                   const nnvm::dim_t nnr_r,
                                   const nnvm::dim_t num_rows,
                                   const nnvm::dim_t num_cols,
-                                  const nnvm::dim_t seg_len,
-                                  const std::unordered_map<RType, nnvm::dim_t>* row_idx_map) {
+                                  const std::unordered_map<RType, nnvm::dim_t>& row_idx_map) {
     using nnvm::dim_t;
-    const dim_t seg_start = i * seg_len;
-    if (seg_start >= num_rows) return;
-    const dim_t seg_end = std::min(seg_start + seg_len, num_rows);
-    for (dim_t j = seg_start; j < seg_end; ++j) {
-      if (indptr_l[j] == indptr_l[j+1]) continue;
-      const dim_t offset_out = j * num_cols;
-      for (IType k = indptr_l[j]; k < indptr_l[j+1]; k++) {
-        for (dim_t l = 0; l < num_cols; ++l) {
-          if (row_idx_map->count(static_cast<RType>(k))) {
-            auto it = row_idx_map->find(static_cast<RType>(k));
-            const dim_t offset_r = it->second * num_cols;
-            out[offset_out+l] += data_l[k] * data_r[offset_r+l];
-          }
-        }
+    if (indptr_l[i] == indptr_l[i+1]) return;
+    const dim_t offset_out = i * num_cols;
+    for (IType k = indptr_l[i]; k < indptr_l[i+1]; k++) {
+      auto it = row_idx_map.find(static_cast<RType>(col_idx_l[k]));
+      if (it == row_idx_map.end()) continue;
+      const dim_t offset_r = it->second * num_cols;
+      for (dim_t l = 0; l < num_cols; ++l) {
+        out[offset_out+l] += data_l[k] * data_r[offset_r+l];
       }
     }
   }
@@ -771,13 +764,13 @@ inline void DotCsrRspDnsImpl(const OpContext& ctx,
             mxnet_op::Kernel<mxnet_op::set_zero, cpu>::Launch(s, num_threads,
                                                               ret->dptr<DType>());
           }
-          num_threads = mxnet_op::get_num_threads<cpu>(ret->shape_[0]);
-          dim_t seg_len = (ret->shape_[0] + num_threads - 1) / num_threads;
+          num_threads = ret->shape_[0];
           if (trans_lhs) {
             LOG(FATAL) << "DotCsrRspDnsImpl has not implemented dot(csr.T, rsp) = dns yet";
           } else {
             const RType* row_idx_ptr = row_idx_r.dptr<RType>();
             std::unordered_map<RType, dim_t> row_idx_map;
+            row_idx_map.reserve(nnr * 19);
             for (dim_t ind = 0; ind < nnr; ind++) {
               row_idx_map.emplace(row_idx_ptr[ind], ind);
             }
@@ -785,7 +778,7 @@ inline void DotCsrRspDnsImpl(const OpContext& ctx,
                 ret->dptr<DType>(), data_l.dptr<DType>(),
                 indptr_l.dptr<IType>(), col_idx_l.dptr<CType>(), data_r.dptr<DType>(),
                 row_idx_r.dptr<RType>(), rhs.storage_shape()[0],
-                ret->shape_[0], ret->shape_[1], seg_len, &row_idx_map);
+                ret->shape_[0], ret->shape_[1], row_idx_map);
           }
         });
       });