@ZiyueHuang 2018-03-08
batch_size = 1000, output_dim = 1, 30 loops

feature dims |density | time (before)  | time (after)  | hash map init
------------ | ------------- | ---------| ---- | ----
 1M | 0.001 | 9.87 | 5.16 | 50% of total
 2M | 0.0005 | 12.19 | 6.20 | 50% of total
 3M | 0.00033    |  13.16| 6.78 | 50% of total
 4M | 0.00025     | 13.65 |    7.08 | 50% of total


batch_size = 1000, output_dim = 1024, 30 loops

feature dims |density | time (before)  | time (after) | hash map init
------------ | ------------- | ---------| ------ | ---
 1M | 0.001 | 16.69 | 15.61 |  12.5% of total
 2M | 0.0005 | 19.46 | 16.09 |  12.5% of total


below is the benchmark script,

```
import mxnet as mx
import numpy as np
import time

mx.random.seed(123)
np.random.seed(123)

out_dims = 1024
batch_size = 1000

for i in range(1, 3):
    fea_dims = 1000000 * i
    density = 0.001 / i
    rshape = (fea_dims, out_dims)
    lshape = (batch_size, fea_dims)
    csr = mx.test_utils.rand_ndarray(lshape, 'csr', density=density)
    dns = mx.test_utils.rand_ndarray(rshape, 'default')
    inds = np.unique(csr.indices.asnumpy())
    print len(inds)
    rsp = mx.nd.sparse.retain(dns.tostype('row_sparse'), mx.nd.array(inds))
    print len(rsp.indices)
    # warm up
    for i in range(5):
        mx.nd.dot(csr, rsp)
    mx.nd.waitall()

    tic = time.time()
    for i in range(30):
        mx.nd.dot(csr, rsp)
    mx.nd.waitall()
    print time.time() - tic
```