@reminisce 2018-03-08
@ZiyueHuang 
1. Could you paste `nnr` for each row in the table? It would be interesting to see how runtime changes with the increase of `nnr`?
2. Could you run the test with the same config used by the two tables in https://github.com/apache/incubator-mxnet/issues/9945, especially the first table, where there are 1000 nnz per row in the csr with random distribution?